moved pages css to css modules
