+++ src/components/Current/Current.module.css
@@ -1,3 +1,3 @@
-.Current {
-    /* css */
+svg {
+    z-index: 99;
 }
