+++ src/components/Marker/Marker.module.css
@@ -1,3 +1,3 @@
-.Marker {
-    /* css */
+svg {
+    z-index: 99;
 }
