--- src/css/home.module.css
@@ -1,28 +1,0 @@
-.mainContainer {
-  display: flex;
-  flex-direction: column;
-  width: 100vw;
-  height: 100vh;
-  background-image: url("../images/texting.jpg");
-  background-position: center; /* Center the image */
-  background-repeat: no-repeat; /* Do not repeat the image */
-  background-size: cover; /* Resize the background image to cover the entire container */
-}
-
-.logoContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 70%;
-}
-
-svg {
-  width: 65%;
-  fill: var(--white);
-}
-
-.slogan {
-  color: var(--white);
-}
