+++ src/pages/Login/index.js
@@ -1,8 +1,8 @@
 import React from "react";
-import Nav from "../components/Nav";
-import LoginForm from "../components/LoginForm";
+import Nav from "../../components/Nav";
+import LoginForm from "../../components/LoginForm";
 
-import "../index.css";
+import "../../index.css";
 
 const Login = props => {
     return (
