+++ src/pages/New/index.js
@@ -1,25 +1,16 @@
 import React, { useState, useEffect } from "react";
-import config from "../config";
-import Location from "../components/Location";
-import Logo from "../components/Logo";
-import Nav from "../components/Nav";
+import Location from "../../components/Location";
+import PostForm from "../../components/PostForm";
+import Logo from "../../components/Logo";
+import Nav from "../../components/Nav";
 
-import io from "socket.io-client";
+import "../../index.css";
 
-import "../index.css";
-
-const socket = io(config.SOC_URL, { transports: ["websocket"] });
-
-const Live = props => {
+const Post = props => {
     const [location, setLocation] = useState({});
-    const [posts, setPosts] = useState([]);
     // const [zoom, setZoom] = useState(props.zoom);
 
     useEffect(() => {
-        socket.on("post", msg => {
-            setPosts([...posts, msg]);
-        });
-
         if (navigator && navigator.geolocation) {
             navigator.geolocation.getCurrentPosition(pos => {
                 setLocation({
@@ -28,22 +19,20 @@ const Live = props => {
                 });
             });
         }
-    }, [posts]);
-
+    }, []);
     return (
-        <div id="live">
+        <div id="new">
             <Logo />
             <Location
-                zoom={12}
+                zoom={11}
                 lat={location.lat}
                 long={location.long}
-                markers={posts}
                 current={true}
                 colour="dodgerblue"
             />
+            <PostForm lat={location.lat} long={location.long} />
             <Nav />
         </div>
     );
 };
-
-export default Live;
+export default Post;
