+++ src/pages/Post/index.js
@@ -1,13 +1,13 @@
 import React, { useState, useEffect } from "react";
 import io from "socket.io-client";
-import config from "../config";
-import Location from "../components/Location";
-import Logo from "../components/Logo";
-import Nav from "../components/Nav";
-import ShowPost from "../components/ShowPost";
-import AddComment from "../components/AddComment";
+import config from "../../config";
+import Location from "../../components/Location";
+import Logo from "../../components/Logo";
+import Nav from "../../components/Nav";
+import ShowPost from "../../components/ShowPost";
+import AddComment from "../../components/AddComment";
 
-import "../index.css";
+import "../../index.css";
 
 const socket = io(config.SOC_URL, { transports: ["websocket"] });
 const ShowPosts = props => {
@@ -52,7 +52,7 @@ const ShowPosts = props => {
         socket.on(`${postId}`, post => {
             setPost([post]);
         });
-    }, [postId]);
+    }, [postId, post]);
 
     return (
         <div id="posts">
