--- src/pages/layout.module.css
@@ -1,6 +1,0 @@
-.mainContainer {
-  display: flex;
-  flex-direction: column;
-  width: 100vw;
-  height: 100vh;
-}
